Move helper functions into utils.go
